Separated the files into subfolders
