+++ lib/main.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
-import 'base_class.dart';
-import 'theme_provider.dart';
-import 'nav_provider.dart';
+import 'widgets/base_class.dart';
+import 'providers/theme_provider.dart';
+import 'providers/nav_provider.dart';
 
 import 'package:hive_flutter/hive_flutter.dart';
 import 'package:provider/provider.dart';
